libs/jk-core/pyproject.toml
[project]
name = "jk-core"
version = "0.1.0"
dependencies = [
    "google-genai>=1.68.0",
    "jinja2>=3.1.6",
    "platformdirs",
    "pyyaml>=6.0.3",
    "numpy",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pyyaml"]

[tool.hatch.build.targets.wheel]
packages = ["src/jk_core"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
